"Dynamic" import of Redis
If Redis is available on the system it is imported while if it is not,
we use a MockRedis class. Redis is used for caching so without we cannot
have caching enabled.

This allows you to use peeringdb on a machine without Redis as long as
caching is disabled while on a machine with Redis, caching can be used
as normally.